Switch back to PathBuf
In preparation for using arc-swap to pass the current file
to the main thread.
(arc-swap doesn't support Arc<?Sized> because that would require
AtomicU128 which is still unstable.
Alternatives would be to use double or triple buffering or my
PairLock crate, but arc-swap sounds like it's faster for my use case,
especially as long as the value is unchanged while main is reading it.