Change DB schema to support file name uniqueness checking
The fields used by queries in the python script are still there
so they don't require any changes.

But this fixes a bug where the data type of printable_path was
BLOB if path was returned (hence why .decode('utf-8') was needed)
but STRING if printable_version was needed,
which would make the script crash due to str not having decode().