Convert existing databases to new schema
SQLites' ALTER TABLE is rather limited in what it can do,
so rename existing, create new, transfer data, delete old is the
reccomended approach.

To be able to create the new schema both from rust and python,
the statements are moved to a sql file.
Documentation of the fields is also improved.